// samples/web-csr/dressca-frontend/admin/src/system-common/generated/api-client/models/put-catalog-item-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Dressca
 * ECサイトDressca
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface PutCatalogItemRequest {
    'catalogBrandId': string;
    'catalogCategoryId': string;
    'description': string;
    'isDeleted': boolean;
    'name': string;
    'price': number;
    'productCode': string;
    'rowVersion': string;
}